Make Jitter Intentionally Non-Async & Recursive Type Discovery & Ready for First PyPi Publish
Non-Async is so that you can arbitrarily run Jitter from within a pre-existing async context (or not!) and it just works. It's actually a *feature* not a bug that Jitter should pause the world and prevent other things from happening "at the same time".

Recursively discovering types means the LLM system context should technically have all the information it needs to correctly implement any function thrown at it. Previously it was blind to the type definition of nested references to user-defined tyeps. Now Jitter auto-extracts types referenced arbitrarily deeply in types in the signature.